Refactor bot into modular components
