+++ bot.py
@@ -1,23 +1,22 @@
 import logging
 import os
-import signal
-import sys
-import asyncio
-import io
-import sqlite3
-import yfinance as yf
-from datetime import datetime
+
 from dotenv import load_dotenv
-from google import genai
-from google.genai import types
+from telegram.ext import Application, CommandHandler
+from telegram.request import HTTPXRequest
 
-from telegram import Update
-from telegram.constants import ChatAction
-from telegram.ext import Application, CommandHandler, CallbackContext
-from telegram.request import HTTPXRequest 
-
-import requests
-from typing import Optional, List, Dict, Tuple
+from database import init_db
+from handlers import (
+    help_command,
+    manage,
+    manage_by_id,
+    open_trade,
+    positions,
+    scan,
+    sentiment,
+    setmodel,
+    start,
+)
 
 load_dotenv()
 
@@ -28,563 +27,15 @@ if not TELEGRAM_TOKEN:
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-# --- DATABASE SETUP (The "Business Ledger" with Expiry) ---
-def init_db():
-    conn = sqlite3.connect('trades.db')
-    c = conn.cursor()
-    # Added 'status' and 'closed_date' to support deterministic position management
-    c.execute('''CREATE TABLE IF NOT EXISTS trades
-                 (id INTEGER PRIMARY KEY AUTOINCREMENT,
-                  chat_id INTEGER,
-                  ticker TEXT,
-                  type TEXT,
-                  strike REAL,
-                  entry_price REAL,
-                  date TEXT,
-                  expiry TEXT,
-                  status TEXT DEFAULT 'OPEN',
-                  closed_date TEXT)''')
 
-    # Backfill schema for existing databases
-    columns = {row[1] for row in c.execute("PRAGMA table_info(trades)")}
-    if 'status' not in columns:
-        c.execute("ALTER TABLE trades ADD COLUMN status TEXT DEFAULT 'OPEN'")
-        c.execute("UPDATE trades SET status = COALESCE(status, 'OPEN')")
-    if 'closed_date' not in columns:
-        c.execute("ALTER TABLE trades ADD COLUMN closed_date TEXT")
+def main() -> None:
+    init_db()
 
-    c.execute("""CREATE INDEX IF NOT EXISTS idx_trades_chat_ticker_status
-                 ON trades (chat_id, ticker, status)""")
-    c.execute("""CREATE INDEX IF NOT EXISTS idx_trades_chat_status
-                 ON trades (chat_id, status)""")
-
-    conn.commit()
-    conn.close()
-
-init_db()
-
-# --- DYNAMIC MODEL TRACKING ---
-user_models = {} 
-
-def resolve_model(chat_id: int, command: str) -> str:
-    """
-    Route to the appropriate model for each command.
-    Gemini handles analysis/scanning with Google Search grounding.
-    Grok is reserved for X/Twitter sentiment.
-    """
-    if command == 'sentiment':
-        return 'grok'
-    if command in ('scan', 'manage', 'manageid'):
-        return 'gemini'
-    return user_models.get(chat_id, 'gemini')
-
-
-def _extract_response_text(response) -> str:
-    """
-    Some Gemini responses (especially tool calls) may not populate .text directly.
-    This helper stitches together any text parts so we always return something user-visible.
-    """
-    if getattr(response, "text", None):
-        return response.text
-
-    text_parts = []
-    for candidate in getattr(response, "candidates", []) or []:
-        content = getattr(candidate, "content", None)
-        if not content:
-            continue
-        for part in getattr(content, "parts", []) or []:
-            part_text = getattr(part, "text", None)
-            if part_text:
-                text_parts.append(part_text)
-    return "\n".join(text_parts).strip()
-
-# --- UPDATED FRAMEWORK CONTEXT ---
-FRAMEWORK_CONTEXT = """
-You are the 'Grandmaster' Trading Assistant. Your core philosophy is 'Be the Casino, Not the Gambler.'
-- Mindset: Sellers collect premiums upfront for an obligation with a statistical edge.
-- Analogy: A credit spread is a 'fence' for a 'dog' (the stock). We only care that the boundary isn't crossed.
-
-The Four Core Trades:
-1. Cash-Secured Puts (CSP): Getting paid to agree to buy the dip.
-2. Covered Calls (CC): Collecting 'rent' on 100+ owned shares.
-3. Bull Put Spreads: Selling a higher-strike put, buying a lower-strike put.
-4. Call Credit Spreads: Selling a lower-strike call, buying a higher-strike call.
-
-Criteria:
-- IV Rank: Favor IV > 50th percentile.
-- Timeframe: Target 30-45 DTE for optimal Theta decay.
-- Management: Close at 50-60% profit. Roll only for a Net Credit.
-"""
-
-HELP_TEXT = """
-🎰 *Hercules "Be the Casino" Tutorial* 🎰
-
-/start - Re-introduces the bot and displays the main command menu.
-/setmodel [model] - Toggle between Grok (best for X-search), OpenAI, and Gemini.
-/scan [ticker] - Analyzes for CSP, CC, BPS, and CCS based on IV and technicals.
-/sentiment [sector or TICKERS] - Sector sentiment or ticker sentiment with auto sector context.
-/manage [ticker] - Checks your trades for 50-60% profit targets or Roll advice.
-/manageid [id] - Manage a specific open trade by its ID.
-/positions [ticker] - List open positions (optionally filtered by ticker).
-/open [ticker] [type] [strike] [premium] [expiry] - Logs your trade (expiry: mm/dd/yyyy).
-
-*Remember: The gold is in managing the position.*
-"""
-
-# --- MARKET DATA HELPERS ---
-def get_market_data(ticker_symbol):
-    try:
-        ticker = yf.Ticker(ticker_symbol)
-        
-        # 1. Robust Price Fetching (Fast Info -> History Fallback)
-        price = "N/A"
-        try:
-            # Try fast_info first (Newer API)
-            price = ticker.fast_info.last_price
-        except:
-            try:
-                # Fallback to history (Most reliable)
-                hist = ticker.history(period="1d")
-                if not hist.empty:
-                    price = hist["Close"].iloc[-1]
-            except Exception as e:
-                logger.warning(f"Price fetch failed for {ticker_symbol}: {e}")
-
-        # Format Price
-        if isinstance(price, (int, float)):
-            price = f"{price:.2f}"
-
-        # 2. Earnings Date (Robust Parsing)
-        next_earnings = "Unknown"
-        try:
-             # Try new earnings_dates DF
-            earnings_df = ticker.earnings_dates
-            if earnings_df is not None and not earnings_df.empty:
-                # FIX: Remove timezone info to allow comparison with naive datetime.now()
-                # This prevents "Invalid comparison between dtype=datetime64[ns, America/New_York] and datetime"
-                future_dates = earnings_df.index[earnings_df.index.tz_localize(None) > datetime.now()]
-                
-                if not future_dates.empty:
-                    next_earnings = future_dates.min().strftime('%Y-%m-%d')
-                else:
-                    # If no future dates, check calendar as last resort
-                    cal = ticker.calendar
-                    if cal and not cal.empty:
-                         next_earnings = cal.iloc[0, 0].strftime('%Y-%m-%d')
-        except Exception as e:
-            logger.warning(f"Earnings fetch failed for {ticker_symbol}: {e}")
-
-        # 3. Get Info (Safely)
-        info = {}
-        try:
-            info = ticker.info
-        except:
-            pass
-
-        return {
-            "price": price,
-            "earnings": next_earnings,
-            "iv_hint": info.get("beta", "N/A"),
-            "sector": info.get("sector", "Unknown")
-        }
-
-    except Exception as e:
-        logger.error(f"⚠️ Market Data Crash for {ticker_symbol}: {e}", exc_info=True)
-        return {"price": "N/A", "earnings": "Check Broker", "iv_hint": "N/A", "sector": "Unknown"}
-
-
-def normalize_tickers(tokens: List[str]) -> List[str]:
-    normalized = []
-    for token in tokens:
-        for part in token.split(','):
-            cleaned = part.strip().upper()
-            if cleaned:
-                normalized.append(cleaned)
-    return normalized
-
-
-def is_ticker_like(token: str) -> bool:
-    token = token.strip().upper().strip(',')
-    if not token:
-        return False
-    cleaned = token.replace('.', '').replace('-', '')
-    return cleaned.isalnum() and token == token.upper() and len(cleaned) <= 6
-
-
-def derive_sectors_for_tickers(tickers: List[str]) -> Dict[str, str]:
-    sector_map: Dict[str, str] = {}
-    for ticker in tickers:
-        data = get_market_data(ticker)
-        sector = data.get("sector") or "Unknown"
-        if sector == "Unknown":
-            logger.info("Sector not found for ticker %s", ticker)
-        sector_map[ticker] = sector
-    return sector_map
-
-
-def build_ticker_sentiment_prompt(tickers: List[str], sector_map: Dict[str, str]) -> str:
-    sectors_lines = "\n".join([f"- {t}: {sector_map.get(t, 'Unknown')}" for t in tickers])
-    unique_sectors = [s for s in dict.fromkeys(sector_map.values())]
-    aggregate = ", ".join(unique_sectors) if unique_sectors else "Unknown"
-    return (
-        f"Tickers analyzed: {', '.join(tickers)}\n\n"
-        f"Derived sectors:\n{sectors_lines}\n\n"
-        f"Aggregate sector exposure: {aggregate}\n\n"
-        "Consider both ticker-specific sentiment and broader sector-level tailwinds/headwinds. "
-        "Describe how the tone differs by ticker/sector and note any contrarian or risk signals shaping psychology."
-    )
-
-# --- AI ROUTING LOGIC ---
-async def call_ai(model: str, prompt: str, system_context: str = FRAMEWORK_CONTEXT, task_type: str = "speed") -> Tuple[str, List[str]]:
-    citations: List[str] = []
-
-    if model == 'grok':
-        from xai_sdk import Client
-        from xai_sdk.chat import user, system
-        from xai_sdk.tools import web_search, code_execution, x_search
-
-        api_key = os.getenv('XAI_API_KEY') or os.getenv('GROK_API_KEY')
-        if not api_key:
-            raise ValueError("XAI_API_KEY or GROK_API_KEY not set.")
-
-        client = Client(api_key=api_key)
-
-        chat_kwargs = {
-            "model": "grok-4-1-fast",
-            "tools": [web_search(), code_execution(), x_search()],
-            "include": ["inline_citations"]
-        }
-
-        chat = client.chat.create(**chat_kwargs)
-        chat.append(system(system_context))
-        chat.append(user(prompt))
-
-        final_response = None
-        content_parts: List[str] = []
-
-        try:
-            for response, chunk in chat.stream():
-                final_response = response
-                chunk_text = getattr(chunk, "content", None)
-                if chunk_text:
-                    content_parts.append(chunk_text)
-        except AttributeError:
-            logger.info("xai_sdk chat object does not support streaming; falling back to sample().")
-            final_response = chat.sample()
-            fallback_content = getattr(final_response, "content", "")
-            if fallback_content:
-                content_parts.append(fallback_content)
-
-        content = "".join(content_parts).strip()
-        if not content and final_response:
-            content = getattr(final_response, "content", "") or ""
-
-        raw_citations = getattr(final_response, "citations", None) or []
-        for entry in raw_citations:
-            url = entry if isinstance(entry, str) else getattr(entry, "url", None)
-            if url:
-                citations.append(url)
-
-        return content, citations
-
-    elif model == 'openai':
-        url = 'https://api.openai.com/v1/chat/completions'
-        headers = {'Authorization': f"Bearer {os.getenv('OPENAI_API_KEY')}", 'Content-Type': 'application/json'}
-        body = {"model": "gpt-4o", "messages": [{"role": "system", "content": system_context}, {"role": "user", "content": prompt}]}
-        content = requests.post(url, headers=headers, json=body).json()['choices'][0]['message']['content']
-        return content, citations
-
-    elif model == 'gemini':
-        try:
-            api_key = os.getenv('GEMINI_API_KEY')
-            if not api_key:
-                raise ValueError("GEMINI_API_KEY not set in environment.")
-
-            client = genai.Client(api_key=api_key)
-
-            # Faster scans use Flash; management defaults to Pro for deeper reasoning.
-            model_name = 'gemini-2.5-pro' if task_type == 'reasoning' else 'gemini-2.5-flash'
-            temperature = 0.2 if task_type == 'reasoning' else 0.7
-
-            config_kwargs = {
-                "system_instruction": system_context,
-                "tools": [types.Tool(google_search=types.GoogleSearch())],
-                "temperature": temperature,
-            }
-
-            if task_type == 'reasoning':
-                thinking_cls = getattr(types, "ThinkingConfig", None)
-                if thinking_cls:
-                    config_kwargs["thinking_config"] = thinking_cls()
-
-            config = types.GenerateContentConfig(**config_kwargs)
-
-            response = client.models.generate_content(
-                model=model_name,
-                contents=prompt,
-                config=config,
-            )
-
-            result_text = _extract_response_text(response)
-            content = result_text or "⚠️ AI Error: Empty response from Gemini."
-            return content, citations
-        except Exception as e:
-            logger.error("Gemini API Error: %s", e)
-            return f"⚠️ AI Error: {str(e)}", citations
-
-# --- COMMAND HANDLERS ---
-
-async def start(update: Update, context: CallbackContext):
-    await update.effective_message.reply_markdown(f"Welcome to **HerculesTradingBot**! 🚀\n\n{HELP_TEXT}")
-
-async def setmodel(update: Update, context: CallbackContext):
-    if not context.args: return await update.effective_message.reply_text('Usage: /setmodel [grok|openai|gemini]')
-    model = context.args[0].lower()
-    if model in ['grok', 'openai', 'gemini']:
-        user_models[update.effective_chat.id] = model
-        await update.effective_message.reply_text(
-            f"✅ Model set to {model}. Note: /sentiment always uses Grok; /scan and /manage use Gemini with Google Search."
-        )
-
-async def scan(update: Update, context: CallbackContext):
-    model = resolve_model(update.effective_chat.id, 'scan')
-    ticker_sym = context.args[0].upper() if context.args else 'SOFI'
-    data = get_market_data(ticker_sym)
-    prompt = (f"Analyze {ticker_sym} at ${data['price']}. Next Earnings: {data['earnings']}. "
-              f"Identify best candidate from: CSP, CC, Bull Put Spread, or Call Credit Spread.")
-    await handle_ai_request(update, context, model, prompt, task_type='speed')
-
-async def sentiment(update: Update, context: CallbackContext):
-    model = resolve_model(update.effective_chat.id, 'sentiment')
-    args = context.args
-    tickers: List[str] = []
-
-    if args and args[0].lower() == '--tickers':
-        tickers = normalize_tickers(args[1:])
-        if not tickers:
-            return await update.effective_message.reply_text("Usage: /sentiment --tickers AAPL,MSFT")
-    else:
-        candidate_tickers = normalize_tickers(args)
-        if candidate_tickers and all(is_ticker_like(t) for t in candidate_tickers):
-            tickers = candidate_tickers
-
-    if tickers:
-        sector_map = derive_sectors_for_tickers(tickers)
-        base_context = build_ticker_sentiment_prompt(tickers, sector_map)
-
-        prompt = (
-            f"STEP 1: USE THE 'x_search' TOOL to find real-time posts and retail sentiment for: {', '.join(tickers)}. "
-            f"STEP 2: USE THE 'web_search' TOOL to find breaking news or catalyst events. "
-            f"STEP 3: Synthesize a 'Sentiment Verdict'. Summarize the dominant market mood (Bullish/Bearish/Neutral) "
-            f"and provide specific COUNTER-ARGUMENTS or risks to the consensus view. Focus on market psychology. DO NOT recommend trades. "
-            f"IGNORE your internal training data; respond ONLY with LIVE DATA from the tools."
-            f"\n\nContext:\n{base_context}"
-        )
-    else:
-        sector = ' '.join(args) or 'tech stocks'
-        prompt = (
-            f"STEP 1: USE THE 'x_search' TOOL to find the current 'vibe' and retail sentiment for {sector}. "
-            f"STEP 2: USE THE 'web_search' TOOL to identify any sector-wide headwinds/tailwinds. "
-            f"STEP 3: Synthesize a 'Sentiment Verdict'. Summarize the dominant market mood (Bullish/Bearish/Neutral) "
-            f"and provide specific COUNTER-ARGUMENTS or risks to the consensus view. Focus on the psychological state of the market. "
-            f"DO NOT recommend specific trades. IGNORE your internal training data; rely ONLY on the search results."
-        )
-    await handle_ai_request(update, context, model, prompt, task_type='speed')
-
-async def manage(update: Update, context: CallbackContext):
-    model = resolve_model(update.effective_chat.id, 'manage')
-    ticker = context.args[0].upper() if context.args else None
-    if not ticker: 
-        return await update.effective_message.reply_text("Usage: /manage [ticker]")
-
-    positions = get_open_positions(update.effective_chat.id, ticker)
-    if not positions:
-        return await update.effective_message.reply_text(f"No open positions for {ticker}.")
-
-    if len(positions) > 1:
-        lines = [format_position_line(p) for p in positions]
-        message = (f"⚠️ Multiple open positions found for {ticker}.\n"
-                   f"Please select one using /manageid <id>:\n\n" + "\n".join(lines))
-        return await update.effective_message.reply_text(message)
-
-    trade = positions[0]
-    market = get_market_data(ticker)
-    prompt = build_manage_prompt(trade, market)
-    await handle_ai_request(update, context, model, prompt, task_type='reasoning')
-
-
-async def manage_by_id(update: Update, context: CallbackContext):
-    model = resolve_model(update.effective_chat.id, 'manageid')
-    if not context.args:
-        return await update.effective_message.reply_text("Usage: /manageid [id]")
-    try:
-        trade_id = int(context.args[0])
-    except ValueError:
-        return await update.effective_message.reply_text("Trade id must be a number.")
-
-    trade = get_trade_by_id(trade_id, update.effective_chat.id)
-    if not trade:
-        return await update.effective_message.reply_text("No open trade found with that ID for this chat.")
-
-    market = get_market_data(trade["ticker"])
-    prompt = build_manage_prompt(trade, market)
-    await handle_ai_request(update, context, model, prompt, task_type='reasoning')
-
-
-async def positions(update: Update, context: CallbackContext):
-    ticker_filter = context.args[0].upper() if context.args else None
-    trades = get_open_positions(update.effective_chat.id, ticker_filter)
-    if not trades:
-        msg = f"No open positions{f' for {ticker_filter}' if ticker_filter else ''}."
-        return await update.effective_message.reply_text(msg)
-
-    header = f"Open positions{f' for {ticker_filter}' if ticker_filter else ''}:"
-    lines = [format_position_line(t) for t in trades]
-    await update.effective_message.reply_text(f"{header}\n" + "\n".join(lines))
-
-async def open_trade(update: Update, context: CallbackContext):
-    """
-    Command: /open [ticker] [type] [strike] [premium] [expiry]
-    """
-    args = context.args
-    arg_count = len(args)
-    
-    # DIAGNOSTIC: If it fails, tell us exactly why
-    if arg_count != 5:
-        error_msg = (
-            f"❌ **Argument Mismatch**\n"
-            f"Expected: 5 arguments\n"
-            f"Received: {arg_count}\n\n"
-            f"Your input: `{args}`\n\n"
-            f"Usage: `/open [TICKER] [TYPE] [STRIKE] [PREMIUM] [MM/DD/YYYY]`"
-        )
-        return await update.effective_message.reply_markdown(error_msg)
-
-    try:
-        ticker, t_type, strike, premium, expiry = args
-        
-        # Validate Date format briefly
-        try:
-            datetime.strptime(expiry, '%m/%d/%Y')
-        except ValueError:
-            return await update.effective_message.reply_text("❌ Date must be in MM/DD/YYYY format.")
-
-        conn = sqlite3.connect('trades.db')
-        c = conn.cursor()
-        c.execute("""INSERT INTO trades (chat_id, ticker, type, strike, entry_price, date, expiry, status, closed_date)
-                     VALUES (?, ?, ?, ?, ?, ?, ?, 'OPEN', NULL)""",
-                  (update.effective_chat.id, ticker.upper(), t_type.upper(), 
-                   float(strike), float(premium), 
-                   datetime.now().strftime('%Y-%m-%d'), expiry))
-        conn.commit()
-        conn.close()
-        
-        await update.effective_message.reply_text(f"✅ Business is open! Logged {ticker.upper()} {t_type.upper()} expiring {expiry}.")
-        
-    except Exception as e:
-        await update.effective_message.reply_text(f"⚠️ Database/System Error: {str(e)}")
-
-async def handle_ai_request(update, context, model, prompt, task_type: str = 'speed'):
-    try:
-        await context.bot.send_chat_action(chat_id=update.effective_chat.id, action=ChatAction.TYPING)
-    except Exception as e:
-        logger.warning(f"Could not send typing action (harmless network issue): {e}")
-
-    try:
-        result, citations = await call_ai(model, prompt, task_type=task_type)
-
-        if not result:
-            result = "⚠️ AI returned no text (Check logs for tool output)."
-
-        if citations:
-            deduped = []
-            for url in citations:
-                if url and url not in deduped:
-                    deduped.append(url)
-            if deduped:
-                sources_block = "\n".join(f"- {url}" for url in deduped)
-                result = f"{result}\n\nSources:\n{sources_block}"
-
-        if len(result) > 4000:
-            buffer = io.BytesIO(result.encode('utf-8'))
-            buffer.name = 'response.txt'
-            await update.effective_message.reply_document(document=buffer, caption='Response is long — sent as file.')
-        else:
-            await update.effective_message.reply_text(result)
-    except Exception as e:
-        logger.error("Bot Reply Error: %s", e)
-        await update.effective_message.reply_text(f"⚠️ System Error: {str(e)}")
-
-
-# --- TRADE HELPERS ---
-def row_to_dict(row: sqlite3.Row) -> Dict:
-    return {
-        "id": row["id"],
-        "ticker": row["ticker"],
-        "type": row["type"],
-        "strike": row["strike"],
-        "entry_price": row["entry_price"],
-        "expiry": row["expiry"],
-        "status": row["status"],
-        "closed_date": row["closed_date"],
-        "date": row["date"],
-    }
-
-
-def get_open_positions(chat_id: int, ticker: Optional[str] = None) -> List[Dict]:
-    conn = sqlite3.connect('trades.db')
-    conn.row_factory = sqlite3.Row
-    c = conn.cursor()
-    if ticker:
-        c.execute("""SELECT id, ticker, type, strike, entry_price, date, expiry, status, closed_date
-                     FROM trades
-                     WHERE ticker=? AND chat_id=? AND status='OPEN'
-                     ORDER BY id DESC""", (ticker, chat_id))
-    else:
-        c.execute("""SELECT id, ticker, type, strike, entry_price, date, expiry, status, closed_date
-                     FROM trades
-                     WHERE chat_id=? AND status='OPEN'
-                     ORDER BY id DESC""", (chat_id,))
-    rows = [row_to_dict(r) for r in c.fetchall()]
-    conn.close()
-    return rows
-
-
-def get_trade_by_id(trade_id: int, chat_id: int) -> Optional[Dict]:
-    conn = sqlite3.connect('trades.db')
-    conn.row_factory = sqlite3.Row
-    c = conn.cursor()
-    c.execute("""SELECT id, ticker, type, strike, entry_price, date, expiry, status, closed_date
-                 FROM trades
-                 WHERE id=? AND chat_id=? AND status='OPEN'
-                 LIMIT 1""", (trade_id, chat_id))
-    row = c.fetchone()
-    conn.close()
-    return row_to_dict(row) if row else None
-
-
-def format_position_line(trade: Dict) -> str:
-    return f"• ID {trade['id']} — {trade['ticker']} {trade['type']} {trade['strike']} exp {trade['expiry']} entry {trade['entry_price']}"
-
-
-def build_manage_prompt(trade: Dict, market: Dict) -> str:
-    # Explicitly include Strike Price to prevent AI hallucinating the Premium as the Strike
-    entry_info = (f"Position: {trade['type']} @ Strike: ${trade['strike']}. "
-                  f"Premium Collected: ${trade['entry_price']}. "
-                  f"Expiry: {trade['expiry']} (Opened: {trade['date']})")
-    
-    return (f"Manage {trade['ticker']}. {entry_info}. Current Market Price: ${market['price']}. "
-            f"Today: {datetime.now().strftime('%Y-%m-%d')}. "
-            f"Calculate current profit/loss based on decay. "
-            f"Evaluate 50% profit target and provide Net Credit Roll advice.")
-
-def main():
     request = HTTPXRequest(connect_timeout=60.0, read_timeout=60.0)
-    
     application = Application.builder().token(TELEGRAM_TOKEN).request(request).build()
-    
+
     application.add_handler(CommandHandler("start", start))
-    application.add_handler(CommandHandler("help", lambda u, c: u.effective_message.reply_markdown(HELP_TEXT)))
+    application.add_handler(CommandHandler("help", help_command))
     application.add_handler(CommandHandler("setmodel", setmodel))
     application.add_handler(CommandHandler("scan", scan))
     application.add_handler(CommandHandler("sentiment", sentiment))
@@ -592,7 +43,10 @@ def main():
     application.add_handler(CommandHandler("manageid", manage_by_id))
     application.add_handler(CommandHandler("positions", positions))
     application.add_handler(CommandHandler("open", open_trade))
-    print("Bot is running...")
+
+    logger.info("Bot is running...")
     application.run_polling()
 
-if __name__ == '__main__': main()
+
+if __name__ == '__main__':
+    main()
